Melhorei a nova implementaçao Alterei o GameResults para que so de os em Jogo Alterei o getProgrammerInfo para que so de os em Jogo Alterei as consequencias Alterei as casas Mudei uma coisa no moveCurrentPlayer para funcionar

+++ src/pt/ulusofona/lp2/deisiGreatGame/Programmer.java
@@ -154,7 +154,7 @@ public class Programmer {
             case "Erro de lógica": {
                 if (!(removeFerramenta(new JUnit()) || removeFerramenta(new AjudaProfessor()))) {
                     try {
-                        andar(-(int) (nrSpaces / 2));
+                        andar(-(nrSpaces / 2));
                     } catch (java.lang.Exception c) {
                         c.printStackTrace();
                     }
